Rewrite history handling to use buffers
The previous system handled this with timeouts. This allowed the code to
be incredible simple but rather inefficient and inflexible. This new
code may not work, but when it does work it will use less memory and
give options later for better data presentation.

The new system creates a two buffers for every template. Each entry in
these buffers represents pixels placed in a given minute. There are two
buffers so negative progress and positive progress can be tracked
separately — as before. These buffers can be summed along a range to
find the number of pixels placed in any given time range — provided the
time range falls withing the buffer's total size. This size defaults to
7 days here but expanding or contracting it should be as easy as
providing a separate value in the constructor.