Update npm libraries
This means we get the joy of updating to the latest unstable Discord.js
version. Obviously, this broke things so those are also fixed here.
Can't wait for them to break again before release. At least it's easy
to see when things break thanks to Typescript — that change is already
paying for itself.