Improve history filling after connection loss
This includes actually resyncing history data if the connection goes
down while the bot is live as well as not resorting to rounding the
average when backfilling.

It also changes the way data is recorded to use a single code-path and
hopefully make things more consistent. That said, the `fillData`
function is far from trivial and it's one of the few places in the code
that I think would benefit significantly from unit tests.