Sync updates with minute epoch
This wasn't as much of an issue before when history data was
millisecond-accurate, since every update would always show the last
minute's data, but — now that data is minute-accurate — updates can be
off by as much as a minute. By syncronising updates to occur just after
the clock change, latency is reduced significantly. Another benefit of
this is that the time of summary updates can be more reliably predicted.