Improve canvas reset behaviour
Since it's still pretty untested, I'm not really sure resets are as
painless as I hope they are. For example, there were several oversights
that are fixed here but there could be more.

Even more concerning is that this debugging was spurred on by a false
reset being tripped — a situation which I was unable to reproduce. Since
this wipes all current data, it would be rather inconvenient if it
happens in production.