Make template command responses ephemeral
This is mostly a preference thing, but I see commands as only needing to
be publically visible when they directly communicate public information
as their purpose. The template management commands currently don't do
that.

It would still be nice to have a command which showcases a template, so
I've noted a TODO for a new `template post` subcommand.