Use node addon for template conversion
The addon — detemplatize — is written in rust. This is much more memory
efficient and faster than both prior implementations.

In testing, I found that processing the global template went from taking
3 seconds at best to taking only about 120 milliseconds. The 3 second
benchmark is for tensorflow — the most recent re-implementation of this
code. Tensorflow had the disadvantage of consuming about 2 gigabytes of
memory to do this. This new module consumes an order of magnitude or two
less memory.

While the implementation of this in node was about as memory efficient
as this, that version could take over 10 seconds to process the same
template now processed in 120 milliseconds.

There's likely still optimization to be done here, but doing so would
likely be wasteful — this is about as fast as this code needs to be.