Add template links to summaries
This is something I held off on for a long time. I had though that I
could use the name of each template's field as a link, but that
apparently don't work. After experimenting a bit with hiding the field
name and emulating it in the field body, the best solution was to just
have a separate link in the body.

This was for two main reasons:
- having a proper field name is a lot nicer semantically
- the title still renders as emnpty space