Flatten tracking hierarchy, remove template command
This removes servers as a concept from the architecture. On one hand,
this makes DMs pretty easy to support. On the other, this concentrates a
lot of code at the root level with little oraganisation.

The hope is that this lack of organisation is temporary and can be
resolved by building simple, sensible abstractions and/or helper
functions.

Functionally, this removes size limits for now but will re-use instances
where it can, making it generally more efficient.

Additionally, this removes the template command and changes summaries to
specify links rather than names. This is made practical thanks to the
reuse of template history and deisgn buffers. It should also simplify
the user experience, making it easier to learn how to use the bot for
the first time.