Use message components to select summaries
This introduces an alternative to specifying message IDs or links which
is a bit more user friendly. Unfortunately, it means that commands now
take place in two parts: the command, then the selection. This is
unfortunate because it means commands are no longer atomic and need to
track state between phases.

This would probably be okay if Discord tracked this state for us, but it
does not. We need to store the arguments provided in the first phase and
somehow retrieve them in the second phase. This is done by associating
each select menu with a custom ID which points back to the original
interaction. The original interaction stores it's arguments in a gloabl
map which can be accessed by the second phase using it's ID.

The code for this is all terribly organised (mainly because it's just
not orgnanised at all).